+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
+- Updated dependencies
+
 ## [0.6.1] - 2023-05-21
-- updated dependencies
+- Updated dependencies
 
 ## [0.6.0] - 2023-03-20
 ### Changed