+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.7.0] - 2023-09-17
 - Updated dependencies
 
 ### Added
@@ -40,7 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking Change**: Compile time verify if ident is given through helper
   trait
 
-[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.6.1...HEAD
+[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/ModProg/attribute-derive/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/ModProg/attribute-derive/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/ModProg/attribute-derive/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/ModProg/attribute-derive/compare/v0.3.1...v0.5.0